+++ toolkit/smart-contracts/commands/src/lib.rs
@@ -10,6 +10,7 @@ pub mod assemble_tx;
 pub mod d_parameter;
 pub mod get_scripts;
 pub mod governance;
+pub mod key_value;
 pub mod permissioned_candidates;
 pub mod register;
 pub mod reserve;
@@ -38,6 +39,9 @@ pub enum SmartContractsCmd {
 	AssembleAndSubmitTx(assemble_tx::AssembleAndSubmitCmd),
 	/// Sign a transaction CBOR using a payment signing key
 	SignTx(sign_tx::SignTxCmd),
+	/// Insert key-value pair into the cardano ledger
+	#[command(subcommand)]
+	KeyValue(key_value::KeyValueCmd),
 }
 
 #[derive(Clone, Debug, clap::Parser)]
@@ -70,6 +74,7 @@ impl SmartContractsCmd {
 			Self::Reserve(cmd) => cmd.execute().await,
 			Self::AssembleAndSubmitTx(cmd) => cmd.execute().await,
 			Self::SignTx(cmd) => cmd.execute().await,
+			Self::KeyValue(cmd) => cmd.execute().await,
 		}?;
 		println!("{}", result);
 		Ok(())