+++ changelog.md
@@ -32,7 +32,7 @@ the feature `pallet-session-compat`.
 * `governance init` when genesis utxo had a script attached, then transaction fee was sometimes calculated incorrectly
 
 ## Added
-
+* `key-value insert` command, that inserts a key-value pair into the key-value store
 * `ariadne_v2` selection algorithm that selects committee respecting D-parameter and candidates
 weights, but has much less variance, thanks to assigning guaranteed seats to candidates with
 expected number of seats greater or equal 1.